Move AsyncBufReadExt methods to AsyncBufRead
