+++ tokio/tests/async_send_sync.rs
@@ -294,8 +294,8 @@ async_assert_fn!(tokio::time::sleep_until(Instant): !Unpin);
 async_assert_fn!(tokio::time::timeout(Duration, BoxFuture<()>): !Unpin);
 async_assert_fn!(tokio::time::timeout_at(Instant, BoxFuture<()>): !Unpin);
 async_assert_fn!(tokio::time::Interval::tick(_): !Unpin);
-async_assert_fn!(tokio::io::AsyncBufReadExt::read_until(&mut BoxAsyncRead, u8, &mut Vec<u8>): !Unpin);
-async_assert_fn!(tokio::io::AsyncBufReadExt::read_line(&mut BoxAsyncRead, &mut String): !Unpin);
+async_assert_fn!(tokio::io::AsyncBufRead::read_until(&mut BoxAsyncRead, u8, &mut Vec<u8>): !Unpin);
+async_assert_fn!(tokio::io::AsyncBufRead::read_line(&mut BoxAsyncRead, &mut String): !Unpin);
 async_assert_fn!(tokio::io::AsyncRead::read(&mut BoxAsyncRead, &mut [u8]): !Unpin);
 async_assert_fn!(tokio::io::AsyncRead::read_exact(&mut BoxAsyncRead, &mut [u8]): !Unpin);
 async_assert_fn!(tokio::io::AsyncRead::read_u8(&mut BoxAsyncRead): !Unpin);
